Wrapped constants in a namespace. Added Arm code using Encoders.
Constants now has a namespace named "Constants" for clarity, allows
shorter naming on the enums. Arm now has a Limits struct to define the
limits for the three arm positions. HandleJointInput uses a timer to
check for 1-3 button presses to set arm joint positions respectively.
Created and defined a bunch of variables in arm.h to make it easier to
change in one place, less digging when a quick change is needed. The
HandleExtensionInput will not work until an encoder is added to the
pulley system on the top of the arm. Ideally, on Monday, I run the code
in DEBUG mode (set MODE value in the initializer list in arm.h to
Mode::DEBUG), manually moving the arm and extension, jotting down the
encoder values to use. Right now, I have just guessed some values,
should still work but might not be what we want. Code builds and looks
right. Also need to make sure the timer methods work as I think they
will for the button press counting.