Comment fix, Added fix for unexpected input.
Needed to add an if statement checking if BPresses is > 0 in front of
where the timer in the Arm files is started. This is because we don't
want B to be held down, the 1s interval to pass, BPresses to be reset,
then the timer to start when the B button is let go, starting the timer
w/out incrementing BPresses. This would cause odd behavior with B input.